fix: matched selectors must add to the props of an element
