+++ docs/index2.html
@@ -13,10 +13,10 @@
         <div class="site-pages-index2-html--469062578-second-page">
             <div class="my-component-100140932-colored">foo=bar</div>
             <div class="with-style-271543454-styled-container">
-                <h2>headline
+                <h2 class="with-style-271543454-h2">headline
                     <p>test</p>
                 </h2>
-                <p>paragraph</p>
+                <p class="with-style-271543454-p">paragraph</p>
                 <div class="my-component--727887898-colored">Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren,
                     no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam
                     et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</div>
