+++ src/templates.js
@@ -8,9 +8,10 @@ const React = require('react');
 const matter = require('gray-matter');
 
 const {Markdown, wrapMarkdown} = require('./markdown');
-const {getMarkup, createScopedCss, getMatchingSelectors} = require('./css');
+const {getMarkup, createScopedCss} = require('./css');
 const {loadHelpers} = require('./helpers');
 const {transformJsx, evaluateHelpers} = require('./jsx');
+const {createElement} = require('./react-create-element');
 
 module.exports = {
 	createTemplates,
@@ -108,7 +109,7 @@ function setupSandbox(templates, sandboxExtras, jsxHelpers, getLocalStyle) {
 				return new Proxy(React, {
 					get: function (target, name) {
 						if (name === 'createElement') {
-							return customCreateElement(proxyTarget);
+							return createElement(proxyTarget);
 						}
 						return target[name];
 					}
@@ -119,104 +120,6 @@ function setupSandbox(templates, sandboxExtras, jsxHelpers, getLocalStyle) {
 	};
 
 	return new Proxy(proxyTarget, proxyHandler);
-}
-
-function customCreateElement(sandbox) {
-	let reverseCssMapping;
-	const getReverseCssMapping = () => {
-		// sandbox.cssMapping is {[classname]: hashed-classname}
-		// this reverses to {[hashed-classname]: classname}
-		if (!reverseCssMapping) {
-			reverseCssMapping = Object.keys(sandbox.cssMapping)
-				.filter(name => !name.includes(' '))
-				.filter(name => name.startsWith('.'))
-				.reduce((reverse, name) => {
-					reverse[sandbox.cssMapping[name]] = name.replace(/^./, '');
-					return reverse;
-				}, {});
-		}
-		return reverseCssMapping;
-	};
-
-	const createElement = React.createElement;
-	return function (...args) {
-		let [tagOrComponent, props, children, ...rest] = args;
-		if (typeof tagOrComponent === 'string') {
-			class DomWrapper extends React.Component {
-				getChildContext() {
-					return {
-						stack: {
-							push: node => {
-								this.domStack = this.domStack || [];
-								this.domStack.push(node);
-							},
-							peek: () => {
-								return [this.context.stack.peek(), this.domStack || []];
-							}
-						}
-					};
-				}
-
-				/**
-				 * @param {any} props
-				 * @param {any} currentNode
-				 */
-				processCssMappings(props, currentNode) {
-					if (sandbox.cssMapping) {
-						const matchingSelectors = getMatchingSelectors(this.context.stack.peek(), Object.keys(sandbox.cssMapping));
-						if (matchingSelectors.length > 0) {
-							this.applyMatchingSelectors(props, matchingSelectors);
-						}
-
-						if (props && props.className) {
-							const className = props.className
-								.split(' ')
-								.map(className => getReverseCssMapping()[className])
-								.join(' ')
-								.trim();
-							if (className) {
-								currentNode.class = className;
-							}
-						}
-					}
-				}
-
-				/**
-				 * Side-effect: Modfies props.
-				 *
-				 * @param {Object} props
-				 * @param {string[]} matchingSelectors
-				 */
-				applyMatchingSelectors(props, matchingSelectors) {
-					if (!props) {
-						props = {};
-					}
-					if (!props.className) {
-						props.className = '';
-					}
-					props.className += matchingSelectors
-						.map(matchingSelector => sandbox.cssMapping[matchingSelector])
-						.join(' ');
-				}
-
-				render() {
-					// note: this calls have side effects - call order matters
-					const currentNode = {tag: tagOrComponent};
-					this.context.stack.push(currentNode);
-					this.processCssMappings(props, currentNode);
-					return createElement.apply(React, [tagOrComponent, props, children, ...rest]);
-				}
-			}
-			DomWrapper.contextTypes = {
-				stack: React.PropTypes.any
-			};
-			DomWrapper.childContextTypes = {
-				stack: React.PropTypes.any
-			};
-			return createElement.apply(React, [DomWrapper, undefined, []]);
-		}
-		return createElement.apply(React, [tagOrComponent, props, children, ...rest]);
-	};
 }
 
 function createLocalStyleFactory(htmlSource, ns, filepath, cssVariables) {
@@ -232,6 +135,11 @@ function contextStackFactory(SFC) {
 	class ContextStack extends React.Component {
 		getChildContext() {
 			return {
+				/*
+				 * css-vars scope
+				 *
+				 * Used to track the variable scope over pages/templates
+				 */
 				scope: {
 					get: () => {
 						if (this.localScope) {
@@ -246,6 +154,11 @@ function contextStackFactory(SFC) {
 						this.context.scope.css(css);
 					}
 				},
+				/*
+				 * dom-stack
+				 *
+				 * Used to track the DOM per page/template to apply css selectors
+				 */
 				stack: {
 					push(node) {
 						this.domeStack = this.domeStack || [];
